Replace merge-options with lodash.merge
On top of not giving me the object mutation I assumed when writing
the original argument parser logic, I took a look at the
underlying source code for merge-options.

It actually uses a property named 'foobar' as part of its
implementation. Like, a *core* part of its implementation.

Contrasting this with the level of attention to detail that I know
goes into Lodash's development leaves me just... beyond words.