Assign text, html, and recipients as .path
Uniformly defering asset loading until after entrypoint processing seems
like a good idea, especially since Nodemailer does most of the heavy
lifting for free under this model.

This also paves the way for `recipients` to be defined as an array, or
some other kind of structure (like maybe a `format` sibling to `path`).